backend/requirements.txt

Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
SQLAlchemy==2.0.31
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
scikit-learn==1.5.0
